更新1.4.9

+++ Netch/Controllers/SSController.cs
@@ -1,6 +1,10 @@
-﻿using System;
+﻿using Netch.Forms;
+using Netch.Utils;
+using System;
 using System.Diagnostics;
 using System.IO;
+using System.Runtime.InteropServices;
+using System.Text;
 using System.Threading;
 
 namespace Netch.Controllers
@@ -25,15 +29,44 @@ namespace Netch.Controllers
         /// <returns>是否启动成功</returns>
         public bool Start(Models.Server server, Models.Mode mode)
         {
+            MainForm.Instance.StatusText($"{Utils.i18N.Translate("Status")}{Utils.i18N.Translate(": ")}{Utils.i18N.Translate("Starting Shadowsocks")}");
+
+            File.Delete("logging\\shadowsocks.log");
+            //从DLL启动Shaowsocks
+            if (Global.Settings.BootShadowsocksFromDLL && (mode.Type == 0 || mode.Type == 1 || mode.Type == 2 || mode.Type == 3))
+            {
+                State = Models.State.Starting;
+                var client = Encoding.UTF8.GetBytes($"0.0.0.0:{Global.Settings.Socks5LocalPort}");
+                var remote = Encoding.UTF8.GetBytes($"{server.Hostname}:{server.Port}");
+                var passwd = Encoding.UTF8.GetBytes($"{server.Password}");
+                var method = Encoding.UTF8.GetBytes($"{server.EncryptMethod}");
+                if (!NativeMethods.Shadowsocks.Info(client, remote, passwd, method))
+                {
+                    State = Models.State.Stopped;
+                    Logging.Info("DLL SS INFO 设置失败！");
+                    return false;
+                }
+                Logging.Info("DLL SS INFO 设置成功！");
+
+                if (!NativeMethods.Shadowsocks.Start())
+                {
+                    State = Models.State.Stopped;
+                    Logging.Info("DLL SS 启动失败！");
+                    return false;
+                }
+                Logging.Info("DLL SS 启动成功！");
+                State = Models.State.Started;
+                return true;
+            }
+
             if (!File.Exists("bin\\Shadowsocks.exe"))
             {
                 return false;
             }
-
             Instance = MainController.GetProcess();
             Instance.StartInfo.FileName = "bin\\Shadowsocks.exe";
 
-            if (!String.IsNullOrWhiteSpace(server.Plugin) && !String.IsNullOrWhiteSpace(server.PluginOption))
+            if (!string.IsNullOrWhiteSpace(server.Plugin) && !string.IsNullOrWhiteSpace(server.PluginOption))
             {
                 Instance.StartInfo.Arguments = $"-s {server.Hostname} -p {server.Port} -b {Global.Settings.LocalAddress} -l {Global.Settings.Socks5LocalPort} -m {server.EncryptMethod} -k \"{server.Password}\" -u --plugin {server.Plugin} --plugin-opts \"{server.PluginOption}\"";
             }
@@ -41,7 +74,7 @@ namespace Netch.Controllers
             {
                 Instance.StartInfo.Arguments = $"-s {server.Hostname} -p {server.Port} -b {Global.Settings.LocalAddress} -l {Global.Settings.Socks5LocalPort} -m {server.EncryptMethod} -k \"{server.Password}\" -u";
             }
-            
+
             if (mode.BypassChina)
             {
                 Instance.StartInfo.Arguments += " --acl default.acl";
@@ -54,7 +87,7 @@ namespace Netch.Controllers
             Instance.Start();
             Instance.BeginOutputReadLine();
             Instance.BeginErrorReadLine();
-            for (int i = 0; i < 1000; i++)
+            for (var i = 0; i < 1000; i++)
             {
                 Thread.Sleep(10);
 
@@ -84,9 +117,16 @@ namespace Netch.Controllers
         {
             try
             {
+                if (Global.Settings.BootShadowsocksFromDLL)
+                {
+                    NativeMethods.Shadowsocks.Stop();
+                    return;
+                }
+
                 if (Instance != null && !Instance.HasExited)
                 {
                     Instance.Kill();
+                    Instance.WaitForExit();
                 }
             }
             catch (Exception e)
@@ -97,25 +137,32 @@ namespace Netch.Controllers
 
         public void OnOutputDataReceived(object sender, DataReceivedEventArgs e)
         {
-            if (!String.IsNullOrWhiteSpace(e.Data))
+            try
             {
-                // File.AppendAllText("logging\\shadowsocks.log", $"{e.Data}\r\n");
-
-                if (State == Models.State.Starting)
+                if (!string.IsNullOrWhiteSpace(e.Data))
                 {
-                    if (Instance.HasExited)
+                    File.AppendAllText("logging\\shadowsocks.log", $"{e.Data}\r\n");
+
+                    if (State == Models.State.Starting)
                     {
-                        State = Models.State.Stopped;
-                    }
-                    else if (e.Data.Contains("listening at"))
-                    {
-                        State = Models.State.Started;
-                    }
-                    else if (e.Data.Contains("Invalid config path") || e.Data.Contains("usage") || e.Data.Contains("plugin service exit unexpectedly"))
-                    {
-                        State = Models.State.Stopped;
+                        if (Instance.HasExited)
+                        {
+                            State = Models.State.Stopped;
+                        }
+                        else if (e.Data.Contains("listening at"))
+                        {
+                            State = Models.State.Started;
+                        }
+                        else if (e.Data.Contains("Invalid config path") || e.Data.Contains("usage") || e.Data.Contains("plugin service exit unexpectedly"))
+                        {
+                            State = Models.State.Stopped;
+                        }
                     }
                 }
+            }
+            catch (Exception ec)
+            {
+                Logging.Info("写入Shadowsocks日志失败" + ec.ToString());
             }
         }
     }
